Fix assemble.h. See Issue #7.
